Add Unsafe.Add<T>(ref T, uint) helper.

+++ src/System.Private.CoreLib/shared/System/Memory.cs
@@ -366,7 +366,7 @@ namespace System
                     }
 #endif
 
-                    refToReturn = ref Unsafe.Add(ref refToReturn, desiredStartIndex);
+                    refToReturn = ref Unsafe.Add(ref refToReturn, (uint)desiredStartIndex);
                     lengthOfUnderlyingSpan = desiredLength;
                 }
